requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
geopy
Pillow
pyTelegramBotAPI
requests
yookassa